fix(chat): isolate debug-log panel, gate listener on debug mode, cap logs (technical review #24)
